+++ bluebottle/projects/admin.py
@@ -5,16 +5,23 @@ from django.utils.translation import ugettext_lazy as _
 from django.contrib import admin
 from django.contrib.admin.sites import NotRegistered
 from django.core.urlresolvers import reverse
+from django.utils.html import escape
 
 from sorl.thumbnail.admin import AdminImageMixin
 
-from bluebottle.projects.models import ProjectBudgetLine, Project, ProjectPhaseLog
-from bluebottle.bb_projects.admin import BaseProjectAdmin, ProjectDocumentInline
+from bluebottle.common.admin_utils import ImprovedModelForm
+from bluebottle.geo.models import Location
+
+from bluebottle.bb_projects.admin import ProjectDocumentInline
 from bluebottle.bb_tasks.admin import TaskAdminInline
 from bluebottle.utils.admin import export_as_csv_action
-
-from .models import PartnerOrganization
+from bluebottle.votes.models import Vote
 from bluebottle.bb_projects.models import ProjectTheme
+
+from geoposition.widgets import GeopositionWidget
+
+
+from .models import PartnerOrganization, ProjectBudgetLine, Project, ProjectPhaseLog
 
 logger = logging.getLogger(__name__)
 
@@ -61,39 +68,67 @@ class ProjectAdminForm(forms.ModelForm):
     theme = forms.ModelChoiceField(queryset=ProjectTheme.objects.all().filter(disabled=False))
 
 
-class ProjectAdmin(BaseProjectAdmin):
+class ProjectAdmin(AdminImageMixin, ImprovedModelForm):
     form = ProjectAdminForm
+    date_hierarchy = 'created'
+    ordering = ('-created',)
+    save_on_top = True
+
+    search_fields = ('title', 'owner__first_name', 'owner__last_name', 'organization__name')
+
+    raw_id_fields = ('owner', 'organization',)
+
+    prepopulated_fields = {'slug': ('title',)}
 
     inlines = (ProjectBudgetLineInline, TaskAdminInline, ProjectDocumentInline, ProjectPhaseLogInline)
 
-    list_filter = BaseProjectAdmin.list_filter + \
-        ('is_campaign', 'theme', 'country__subregion__region', 'partner_organization', FundingFilter)
-    list_display = BaseProjectAdmin.list_display + \
-        ('is_campaign', 'deadline', 'donated_percentage', 'vote_count')
-    list_editable = ('is_campaign', )
+    def get_list_filter(self, request):
+        filters = ('status', 'is_campaign', 'theme', 'country__subregion__region',
+                'partner_organization', FundingFilter)
+        return filters
+
+    def get_list_display(self, request):
+        fields = ('get_title_display', 'get_owner_display', 'created',
+                  'status', 'is_campaign', 'deadline', 'donated_percentage')
+        # Only show Location column if there are any
+        if Location.objects.count():
+            fields +=  ('location', )
+        # Only show Vote_count column if there are any votes
+        if Vote.objects.count():
+            fields +=  ('vote_count', )
+        return fields
+
+    def get_list_editable(self, request):
+        return ('is_campaign', )
 
     readonly_fields = ('owner_link', 'organization_link', 'vote_count',
                        'amount_donated', 'amount_needed', 'popularity')
 
     export_fields = ['title', 'owner', 'created', 'status',
                      'deadline', 'amount_asked', 'amount_donated']
+
     actions = (export_as_csv_action(fields=export_fields), )
 
     fieldsets = (
-        (_('Main'), {'fields': ('owner', 'organization', 'status', 'title',
-                                'slug', 'pitch',
-                                'theme', 'favorite', 'deadline', 'image',
-                                'video_url',
-                                'country', 'language',
+        (_('Main'), {'fields': ('owner', 'owner_link',
+                                'organization', 'organization_link',
                                 'partner_organization',
-                                'latitude', 'longitude',
-                                'amount_asked', 'amount_extra',
-                                'reach', 'is_campaign', 'skip_monthly',
-                                'allow_overfunding', 'story',
-                                'date_submitted', 'campaign_started',
-                                'campaign_ended', 'campaign_funded', 'tags',
-                                'amount_donated', 'amount_needed',
-                                'popularity', 'voting_deadline', 'vote_count')}),
+                                'status', 'title', 'slug', 'is_campaign')}),
+
+        (_('Story'), {'fields': ('pitch', 'story', 'reach')}),
+
+        (_('Details'), {'fields': ('language', 'theme', 'image',
+                                   'video_url', 'country',
+                                   'latitude', 'longitude',
+                                   'location', 'place', 'tags')}),
+
+        (_('Goal'), {'fields': ('amount_asked', 'amount_extra',
+                                'amount_donated','amount_needed',
+                                'popularity', 'vote_count')}),
+
+        (_('Dates'), {'fields': ('deadline', 'date_submitted',
+                                 'campaign_started', 'campaign_ended',
+                                 'campaign_funded', 'voting_deadline')}),
 
         (_('Bank details'), {'fields': ('account_holder_name',
                                         'account_holder_address',
@@ -134,11 +169,36 @@ class ProjectAdmin(BaseProjectAdmin):
         percentage = "%.2f" % (100 * obj.amount_donated / obj.amount_asked)
         return "{0} %".format(percentage)
 
+    def queryset(self, request):
+        # Optimization: Select related fields that are used in admin specific display fields.
+        queryset = super(ProjectAdmin, self).queryset(request)
+        return queryset.select_related('projectpitch', 'projectplan', 'projectcampaign', 'owner',
+                                       'organization')
 
-# We wrapped this in a try because sometimes Project hasn't got registered
-# before it hits this.
-try:
-    admin.site.unregister(Project)
-except NotRegistered:
-    pass
+    def get_title_display(self, obj):
+        if len(obj.title) > 50:
+            return u'<span title="{title}">{short_title} [...]</span>'.format(title=escape(obj.title),
+                                                                              short_title=obj.title[:45])
+        return obj.title
+
+    get_title_display.allow_tags = True
+    get_title_display.admin_order_field = 'title'
+    get_title_display.short_description = _('title')
+
+    def get_owner_display(self, obj):
+        owner_name = obj.owner.get_full_name()
+        if owner_name:
+            owner_name = u' ({name})'.format(name=owner_name)
+        return u'{email}{name}'.format(name=owner_name, email=obj.owner.email)
+
+    get_owner_display.admin_order_field = 'owner__last_name'
+    get_owner_display.short_description = _('owner')
+
+    def project_owner(self, obj):
+        object = obj.owner
+        url = reverse('admin:{0}_{1}_change'.format(object._meta.app_label, object._meta.module_name), args=[object.id])
+        return "<a href='{0}'>{1}</a>".format(str(url), object.first_name + ' ' + object.last_name)
+
+    project_owner.allow_tags = True
+
 admin.site.register(Project, ProjectAdmin)
